Implement additional broker operations
This PR implements most of the broker operations missing in
charm-microceph (compared to what's in charms.ceph). The only
exceptions so far are those related to cache-tiering (because
they need additional classes from charmhelpers). Additionally,
most of the error checking stuff has been removed and delegated
to the external calls. Note that most of the operations that
call out to 'radosgw-admin' will not work unless RadosGW is
installed, but they can be placeholders until we know what
to do with these.

Signed-off-by: Luciano Lo Giudice <luciano.logiudice@canonical.com>